# docs/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API documentation generated from YAML routing definitions",
    "title": "GOTRS API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/agent/customers/{id}/tickets": {
      "get": {
        "description": "Get all tickets for a specific customer",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:read",
              "ticket:read"
            ]
          }
        ],
        "summary": "Customer Tickets",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/customer/company/users": {
      "get": {
        "description": "View other users in the same company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "company:users:read"
            ]
          }
        ],
        "summary": "Company Users",
        "tags": [
          "customer"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:8080"
    }
  ]
}